Another round of ports (#1183)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Hair Ports:
- https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/1340
- https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/1372
- https://github.com/Rotwood-Vale/Ratwood-Keep/pull/1520
- https://github.com/Darkrp-community/OpenKeep/pull/1275
- https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/1117
- https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/1529

Azure Ports:
- https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/1483
- https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/1343
- https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/1481
- https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/1456
- https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/1742

Changes:
- Zigboxes have a defined storage size
- Reduces coin counting message range
- Adds salvage results to more clothes
- Fixes bread sprite
- Bottle bomb needs alcohol to craft
- Attack arrows and bolts with fyritius to make pyroclastic variants
- No more guaranteed throw range on mobs, tackle throw-down requires two
aggro grabs instead of one

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Was supposed to just be ports but added too many of my own changes.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

+++ code/modules/food_and_drinks/food/snacks.dm
@@ -140,7 +140,7 @@ All foods are distributed among various categories. Use common sense.
 	if(rotprocess)
 		var/obj/structure/closet/crate/chest/chest = locate(/obj/structure/closet/crate/chest) in get_turf(src)
 		var/obj/structure/fake_machine/vendor = locate(/obj/structure/fake_machine/vendor) in get_turf(src)
-		if(!chest && !vendor)
+		if(!chest && !vendor && !istype(loc, /obj/item/storage/backpack/backpack/artibackpack))
 			var/obj/structure/table/located = locate(/obj/structure/table) in loc
 			if(located)
 				warming -= 5
@@ -319,6 +319,7 @@ All foods are distributed among various categories. Use common sense.
 			mob_location.put_in_hands(generate_trash(mob_location))
 		else
 			generate_trash(current_loc.drop_location())
+	update_icon()
 
 /obj/item/reagent_containers/food/snacks/attack_self(mob/user)
 	return
@@ -408,7 +409,7 @@ All foods are distributed among various categories. Use common sense.
 				bitecount++
 				on_consume(M)
 				checkLiked(fraction, M)
-				if(bitecount >= bitesize)
+				if(bitecount >= bitesize && !QDELETED(src))
 					qdel(src)
 				return TRUE
 		playsound(M.loc,'sound/misc/eat.ogg', rand(30,60), TRUE)
@@ -618,7 +619,7 @@ All foods are distributed among various categories. Use common sense.
 	if(eater.dropItemToGround(src))
 		qdel(src)
 	var/obj/item/I = new path(T)
-	eater.put_in_active_hand(I)
+	eater.put_in_active_hand(I, ignore_animation = TRUE)
 
 /obj/item/reagent_containers/food/snacks/Destroy()
 	STOP_PROCESSING(SSobj, src)
@@ -691,9 +692,9 @@ All foods are distributed among various categories. Use common sense.
 	else
 		return ..()
 
-/obj/item/reagent_containers/food/snacks/on_consume(mob/living/eater)
-	..()
-	if(biting)
+/obj/item/reagent_containers/food/snacks/update_icon()
+	. = ..()
+	if(biting && bitecount)
 		icon_state = "[base_icon_state][bitecount]"
